backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

python-dotenv==1.1.1
python-multipart==0.0.20
email-validator==2.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.5.0

pydantic==2.12.4
pydantic-settings==2.12.0

requests==2.32.5
fastapi-mail==1.5.8
itsdangerous==2.2.0
redis==5.0.4